+++ src/Decred/Signer.cpp
@@ -173,7 +173,7 @@ Data Signer::pushAll(const std::vector<Data>& results) {
             encode16LE(result.size(), data);
         } else {
             data.push_back(OP_PUSHDATA4);
-            encode32LE(result.size(), data);
+            encode32LE(static_cast<uint32_t>(result.size()), data);
         }
         std::copy(begin(result), end(result), back_inserter(data));
     }
