+++ src/interface/TWBinanceSigner.cpp
@@ -17,7 +17,7 @@ using namespace TW::Binance;
 
 TW_Binance_Proto_SigningOutput TWBinanceSignerSign(TW_Binance_Proto_SigningInput data) {
     Proto::SigningInput input;
-    input.ParseFromArray(TWDataBytes(data), TWDataSize(data));
+    input.ParseFromArray(TWDataBytes(data), static_cast<int>(TWDataSize(data)));
 
     auto signer = new TWBinanceSigner{ Signer(std::move(input)) };
     auto encoded = signer->impl.build();
